Keep double precision off the compute device throughout the alignment package
The rotation function's back half accumulated one step wider than the
expansion -- complex128 for the radial sum, the Wigner contraction and the
FFT -- and its inputs, the dense P1 transform, the shell sums and the
expansion's clustering keys were cast to float64 on the device. None of it
survives on a backend without float64, and none of it is needed: measured
on the pose panel, single precision everywhere on the device recovers 30/30
at the default window and 30/30 uncut, with per-alignment times unchanged
(0.44-1.12 s at depth 10, job 551437). The earlier measurement that
motivated the wide accumulator showed the deep peak list reordering while
the top peak stayed put; the placement search now consumes only the top
few distinct orientations.

The expansion's clustering keys need double -- _GROUP_SCALE_S keys |s| at
1e-7, below float32's resolution -- and are formed on the host, which
always has it; only the integer keys reach the device. The Bessel ladder
runs in its argument's dtype, kept in range by its power-of-two rescaling,
so the bit-identity and scipy checks still exercise a double ladder. What
remains in double is host-side: 3x3 rotation algebra, the Wigner-d
eigendecomposition, the anisotropy fit and RotationSolutions.

The two tests that pinned the wider accumulator now pin the configured
dtype. 227 alignment, scaling and Bessel tests (552234); full unit gate
2043 passed (552236); panel 30/30 at both windows (552235).

+++ torchref/experimental/alignment/sh.py
@@ -50,8 +50,9 @@ def legendre_recurrence_coefficients(L: int, dtype, device):
     runs this recurrence itself, fused with its own accumulation, and two copies
     of these formulae would be two chances to get them subtly different.
     """
-    ll = torch.arange(L, dtype=torch.float64, device=device).view(L, 1)  # dtype-ok: harmonic indices as exact doubles
-    mm = torch.arange(L, dtype=torch.float64, device=device).view(1, L)  # dtype-ok: harmonic indices as exact doubles
+    # Small integers, exact in any float dtype; the results are cast to `dtype`.
+    ll = torch.arange(L, dtype=dtype, device=device).view(L, 1)
+    mm = torch.arange(L, dtype=dtype, device=device).view(1, L)
     valid = ll > mm
     denom = (ll - mm) * (ll + mm)
     denom_safe = torch.where(valid, denom, torch.ones_like(denom))
@@ -62,7 +63,7 @@ def legendre_recurrence_coefficients(L: int, dtype, device):
         b_num / torch.where(b_den == 0, torch.ones_like(b_den), b_den), min=0.0))
     a = torch.where(valid, a, torch.zeros_like(a)).to(dtype)
     b = torch.where(valid, b, torch.zeros_like(b)).to(dtype)
-    m_arange = torch.arange(L, dtype=torch.float64, device=device)  # dtype-ok: harmonic indices as exact doubles
+    m_arange = torch.arange(L, dtype=dtype, device=device)
     sect = torch.sqrt(
         (2.0 * m_arange + 1.0) / (2.0 * m_arange).clamp(min=1.0)).to(dtype)
     return a, b, sect
@@ -241,7 +242,7 @@ def equal_count_shell_edges(
     s_sorted, _ = torch.sort(s)
     N = s_sorted.numel()
     # quantile-based partition
-    idx = torch.linspace(0, N - 1, P + 1, dtype=torch.float64, device=s.device).round().long()  # dtype-ok: equal-count edge positions in double, then rounded
+    idx = torch.linspace(0, N - 1, P + 1, dtype=s.dtype, device=s.device).round().long()
     edges = s_sorted[idx]
     # nudge endpoints so the data is fully covered (avoid floating-point miss)
     if s_min is not None:
